# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorSpec": {
        "properties": {
          "affiliation": {
            "default": "community",
            "enum": [
              "ltx",
              "community"
            ],
            "title": "Affiliation",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name"
        ],
        "title": "AuthorSpec",
        "type": "object"
      },
...
      "CatalogDownloadStartResponse": {
        "properties": {
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "sessionId"
        ],
        "title": "CatalogDownloadStartResponse",
        "type": "object"
      },
...
      "DownloadSpec": {
        "description": "HF download location for a catalog item.\n\n``variants`` is always non-empty. The first entry is the default checkpoint\n(preferred for generation / omit-``variant_id`` downloads). See the comment on\n``LoraCatalogHandler`` listing helpers.",
        "properties": {
          "repo_id": {
            "title": "Repo Id",
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/DownloadVariant"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "repo_id",
          "variants"
        ],
        "title": "DownloadSpec",
        "type": "object"
      },
      "DownloadVariant": {
        "description": "One downloadable weights file under a catalog item (e.g. strong vs light).",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "filename",
          "size_bytes"
        ],
        "title": "DownloadVariant",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "IcLoraCatalogItem": {
        "description": "IC-LoRA — base plus IC-specific preprocessing / controls / default settings.\n\nIC-LoRAs always take a driving image/video; `input` is required in practice (the base\nfield is optional for plain LoRAs). Enforced by the validator below, so handlers can rely\non it being present.",
        "properties": {
          "allows_empty_prompt": {
            "default": false,
            "title": "Allows Empty Prompt",
            "type": "boolean"
          },
          "allows_reference_image": {
            "default": false,
            "title": "Allows Reference Image",
            "type": "boolean"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model"
          },
          "controls": {
            "items": {
              "$ref": "#/components/schemas/IcLoraControl"
            },
            "title": "Controls",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "default_settings": {
            "$ref": "#/components/schemas/IcLoraSettings"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "download": {
            "$ref": "#/components/schemas/DownloadSpec"
          },
          "enhancement_examples": {
            "items": {
              "type": "string"
            },
            "title": "Enhancement Examples",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InputSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "instructions": {
            "items": {
              "$ref": "#/components/schemas/InstructionSection"
            },
            "title": "Instructions",
            "type": "array"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preprocessing": {
            "items": {
              "$ref": "#/components/schemas/PreprocessingStep"
            },
            "title": "Preprocessing",
            "type": "array"
          },
          "prompt_template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptTemplateSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommended_strength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Strength"
          },
          "requires_hf_login": {
            "title": "Requires Hf Login",
            "type": "boolean"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger"
          },
          "trigger_placement": {
            "anyOf": [
              {
                "enum": [
                  "first_token",
                  "anywhere"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Placement"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "download",
          "requires_hf_login"
        ],
        "title": "IcLoraCatalogItem",
        "type": "object"
      },
      "IcLoraControl": {
        "properties": {
          "default": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "int",
            "enum": [
              "int",
              "select",
              "position_canvas"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_labels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Labels"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "IcLoraControl",
        "type": "object"
      },
      "IcLoraDownloadProgressResponse": {
        "properties": {
          "downloaded_bytes": {
            "default": 0,
            "title": "Downloaded Bytes",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expected_bytes": {
            "default": 0,
            "title": "Expected Bytes",
            "type": "integer"
          },
          "ic_lora_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ic Lora Id"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "speed_bytes_per_sec": {
            "default": 0.0,
            "title": "Speed Bytes Per Sec",
            "type": "number"
          },
          "status": {
            "enum": [
              "downloading",
              "complete",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "IcLoraDownloadProgressResponse",
        "type": "object"
      },
      "IcLoraDownloadRequest": {
        "properties": {
          "ic_lora_id": {
            "title": "Ic Lora Id",
            "type": "string"
          },
          "use_hf_auth": {
            "default": false,
            "title": "Use Hf Auth",
            "type": "boolean"
          },
          "variant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          }
        },
        "required": [
          "ic_lora_id"
        ],
        "title": "IcLoraDownloadRequest",
        "type": "object"
      },
...
      "IcLoraListItem": {
        "properties": {
          "downloaded": {
            "title": "Downloaded",
            "type": "boolean"
          },
          "downloaded_variant_ids": {
            "items": {
              "type": "string"
            },
            "title": "Downloaded Variant Ids",
            "type": "array"
          },
          "ic_lora": {
            "$ref": "#/components/schemas/IcLoraCatalogItem"
          }
        },
        "required": [
          "ic_lora",
          "downloaded"
        ],
        "title": "IcLoraListItem",
        "type": "object"
      },
      "IcLoraListResponse": {
        "properties": {
          "ic_loras": {
            "items": {
              "$ref": "#/components/schemas/IcLoraListItem"
            },
            "title": "Ic Loras",
            "type": "array"
          }
        },
        "required": [
          "ic_loras"
        ],
        "title": "IcLoraListResponse",
        "type": "object"
      },
      "IcLoraSettings": {
        "properties": {
          "audio_mode": {
            "default": "generated",
            "enum": [
              "source",
              "generated",
              "off"
            ],
            "title": "Audio Mode",
            "type": "string"
          },
          "conditioning_strength": {
            "default": 1.0,
            "title": "Conditioning Strength",
            "type": "number"
          },
          "lora_strength": {
            "default": 1.0,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Lora Strength",
            "type": "number"
          },
          "resolution_factor": {
            "default": 2.0,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Resolution Factor",
            "type": "number"
          },
          "skip_stage_2": {
            "default": false,
            "title": "Skip Stage 2",
            "type": "boolean"
          },
          "use_lora_in_stage_2": {
            "default": false,
            "title": "Use Lora In Stage 2",
            "type": "boolean"
          }
        },
        "title": "IcLoraSettings",
        "type": "object"
      },
...
      "InputSpec": {
        "properties": {
          "kind": {
            "enum": [
              "image",
              "video"
            ],
            "title": "Kind",
            "type": "string"
          },
          "optional": {
            "default": false,
            "title": "Optional",
            "type": "boolean"
          }
        },
        "required": [
          "kind"
        ],
        "title": "InputSpec",
        "type": "object"
      },
...
      "InstructionSection": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Body"
          },
          "kind": {
            "enum": [
              "summary",
              "prompting",
              "tips",
              "input"
            ],
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "enum": [
              "What it does",
              "Input",
              "Prompt",
              "Tips",
              "Notes"
            ],
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title",
          "body"
        ],
        "title": "InstructionSection",
        "type": "object"
      },
      "JsonValue": {},
...
      "LicenseSpec": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name"
        ],
        "title": "LicenseSpec",
        "type": "object"
      },
      "LoraCatalogItem": {
        "description": "Base catalog entry — used as-is for a plain LoRA.",
        "properties": {
          "allows_empty_prompt": {
            "default": false,
            "title": "Allows Empty Prompt",
            "type": "boolean"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "download": {
            "$ref": "#/components/schemas/DownloadSpec"
          },
          "enhancement_examples": {
            "items": {
              "type": "string"
            },
            "title": "Enhancement Examples",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InputSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "instructions": {
            "items": {
              "$ref": "#/components/schemas/InstructionSection"
            },
            "title": "Instructions",
            "type": "array"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt_template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptTemplateSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommended_strength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Strength"
          },
          "requires_hf_login": {
            "title": "Requires Hf Login",
            "type": "boolean"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger"
          },
          "trigger_placement": {
            "anyOf": [
              {
                "enum": [
                  "first_token",
                  "anywhere"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Placement"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "download",
          "requires_hf_login"
        ],
        "title": "LoraCatalogItem",
        "type": "object"
      },
      "LoraDownloadProgressResponse": {
        "properties": {
          "downloaded_bytes": {
            "default": 0,
            "title": "Downloaded Bytes",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expected_bytes": {
            "default": 0,
            "title": "Expected Bytes",
            "type": "integer"
          },
          "lora_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lora Id"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "speed_bytes_per_sec": {
            "default": 0.0,
            "title": "Speed Bytes Per Sec",
            "type": "number"
          },
          "status": {
            "enum": [
              "downloading",
              "complete",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "LoraDownloadProgressResponse",
        "type": "object"
      },
      "LoraDownloadRequest": {
        "properties": {
          "lora_id": {
            "title": "Lora Id",
            "type": "string"
          },
          "use_hf_auth": {
            "default": false,
            "title": "Use Hf Auth",
            "type": "boolean"
          },
          "variant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          }
        },
        "required": [
          "lora_id"
        ],
        "title": "LoraDownloadRequest",
        "type": "object"
      },
...
      "LoraListItem": {
        "properties": {
          "downloaded": {
            "title": "Downloaded",
            "type": "boolean"
          },
          "downloaded_variant_ids": {
            "items": {
              "type": "string"
            },
            "title": "Downloaded Variant Ids",
            "type": "array"
          },
          "lora": {
            "$ref": "#/components/schemas/LoraCatalogItem"
          }
        },
        "required": [
          "lora",
          "downloaded"
        ],
        "title": "LoraListItem",
        "type": "object"
      },
      "LoraListResponse": {
        "properties": {
          "loras": {
            "items": {
              "$ref": "#/components/schemas/LoraListItem"
            },
            "title": "Loras",
            "type": "array"
          }
        },
        "required": [
          "loras"
        ],
        "title": "LoraListResponse",
        "type": "object"
      },
...
      "MediaSpec": {
        "properties": {
          "demo_video": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Video"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          }
        },
        "title": "MediaSpec",
        "type": "object"
      },
...
      "PreprocessingStep": {
        "properties": {
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Params",
            "type": "object"
          },
          "utility": {
            "title": "Utility",
            "type": "string"
          }
        },
        "required": [
          "utility"
        ],
        "title": "PreprocessingStep",
        "type": "object"
      },
...
      "PromptTemplateSpec": {
        "description": "A fixed prompt structure a LoRA/IC-LoRA requires verbatim (e.g. a two-part\n\"Reference shows X. Edited shows Y.\" restore template, or crossview's enum-constrained\ncamera vocabulary). `template` contains `{name}` placeholders; each must have a matching\nentry in `placeholders`. A degenerate template with no placeholders (e.g. \"upscale\") is a\nfixed, non-generated prompt.",
        "properties": {
          "placeholders": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PromptTemplatePlaceholder"
            },
            "title": "Placeholders",
            "type": "object"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "template"
        ],
        "title": "PromptTemplateSpec",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LTX-2 Video Generation Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ic-loras": {
      "get": {
        "operationId": "route_list_ic_loras_api_ic_loras_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcLoraListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Ic Loras",
        "tags": [
          "ic-loras"
        ]
      }
    },
    "/api/ic-loras/download": {
      "post": {
        "operationId": "route_ic_lora_download_api_ic_loras_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IcLoraDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogDownloadStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Ic Lora Download",
        "tags": [
          "ic-loras"
        ]
      }
    },
    "/api/ic-loras/download/progress": {
      "get": {
        "operationId": "route_ic_lora_download_progress_api_ic_loras_download_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcLoraDownloadProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Ic Lora Download Progress",
        "tags": [
          "ic-loras"
        ]
      }
    },
    "/api/loras": {
      "get": {
        "operationId": "route_list_loras_api_loras_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoraListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Loras",
        "tags": [
          "loras"
        ]
      }
    },
    "/api/loras/download": {
      "post": {
        "operationId": "route_lora_download_api_loras_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoraDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogDownloadStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Lora Download",
        "tags": [
          "loras"
        ]
      }
    },
    "/api/loras/download/progress": {
      "get": {
        "operationId": "route_lora_download_progress_api_loras_download_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoraDownloadProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Lora Download Progress",
        "tags": [
          "loras"
        ]
      }
    },
...
  }
}